+++ renderer/skins/default.py
@@ -2,7 +2,8 @@ import json
 from math import ceil
 from pathlib import Path
 
-from renderer.types import CS, CTI, Coord, SkinBuilder, _darken, _lighten
+from renderer.types import CS, CTI, SkinBuilder, _darken, _lighten
+from renderer.types.coord import WorldCoord
 
 A_ROAD = 0xFFAAAA
 B_ROAD = 0xFF8000
@@ -573,7 +574,7 @@ def main():
     pedestrian_crossing[0:5] = [
         CS.point_image(file=Path("pedestriancrossing.png")),
         CS.point_text(colour=0x000000,
-                      offset=Coord(0, 32),
+                      offset=WorldCoord(0, 32),
                       size=32,
                       anchor="mm")
     ]
@@ -583,7 +584,7 @@ def main():
     rail_crossing[0:5] = [
         CS.point_image(file=Path("pedestriancrossing.png")),
         CS.point_text(colour=0x000000,
-                      offset=Coord(0, 32),
+                      offset=WorldCoord(0, 32),
                       size=32,
                       anchor="mm")
     ]
@@ -593,7 +594,7 @@ def main():
     parking[0:5] = [
         CS.point_image(file=Path("parking.png")),
         CS.point_text(colour=0x0000dc,
-                      offset=Coord(0, 32),
+                      offset=WorldCoord(0, 32),
                       size=32,
                       anchor="mm")
     ]
@@ -603,7 +604,7 @@ def main():
     bike_rack[0:5] = [
         CS.point_image(file=Path("bikerack.png")),
         CS.point_text(colour=0x000000,
-                      offset=Coord(0, 32),
+                      offset=WorldCoord(0, 32),
                       size=32,
                       anchor="mm")
     ]
@@ -613,7 +614,7 @@ def main():
     shop[0:5] = [
         CS.point_image(file=Path("shop.png")),
         CS.point_text(colour=0x000000,
-                      offset=Coord(0, 32),
+                      offset=WorldCoord(0, 32),
                       size=32,
                       anchor="mm")
     ]
@@ -623,7 +624,7 @@ def main():
     restaurant[0:5] = [
         CS.point_image(file=Path("restaurant.png")),
         CS.point_text(colour=0x000000,
-                      offset=Coord(0, 32),
+                      offset=WorldCoord(0, 32),
                       size=32,
                       anchor="mm")
     ]
@@ -633,7 +634,7 @@ def main():
     hotel[0:5] = [
         CS.point_image(file=Path("hotel.png")),
         CS.point_text(colour=0x0000fc,
-                      offset=Coord(0, 32),
+                      offset=WorldCoord(0, 32),
                       size=32,
                       anchor="mm")
     ]
@@ -643,7 +644,7 @@ def main():
     arcade[0:5] = [
         CS.point_image(file=Path("arcade.png")),
         CS.point_text(colour=0xffdc00,
-                      offset=Coord(0, 32),
+                      offset=WorldCoord(0, 32),
                       size=32,
                       anchor="mm")
     ]
@@ -653,7 +654,7 @@ def main():
     supermarket[0:5] = [
         CS.point_image(file=Path("supermarket.png")),
         CS.point_text(colour=0x000000,
-                      offset=Coord(0, 32),
+                      offset=WorldCoord(0, 32),
                       size=32,
                       anchor="mm")
     ]
@@ -663,7 +664,7 @@ def main():
     clinic[0:5] = [
         CS.point_image(file=Path("clinic.png")),
         CS.point_text(colour=0xff0303,
-                      offset=Coord(0, 32),
+                      offset=WorldCoord(0, 32),
                       size=32,
                       anchor="mm")
     ]
@@ -673,7 +674,7 @@ def main():
     library[0:5] = [
         CS.point_image(file=Path("library.png")),
         CS.point_text(colour=0x000000,
-                      offset=Coord(0, 32),
+                      offset=WorldCoord(0, 32),
                       size=32,
                       anchor="mm")
     ]
@@ -683,7 +684,7 @@ def main():
     place_of_worship[0:5] = [
         CS.point_image(file=Path("placeofworship.png")),
         CS.point_text(colour=0x000000,
-                      offset=Coord(0, 32),
+                      offset=WorldCoord(0, 32),
                       size=32,
                       anchor="mm")
     ]
@@ -693,7 +694,7 @@ def main():
     petrol[0:5] = [
         CS.point_image(file=Path("petrol.png")),
         CS.point_text(colour=0x000000,
-                      offset=Coord(0, 32),
+                      offset=WorldCoord(0, 32),
                       size=32,
                       anchor="mm")
     ]
@@ -703,7 +704,7 @@ def main():
     cinema[0:5] = [
         CS.point_image(file=Path("cinema.png")),
         CS.point_text(colour=0x000000,
-                      offset=Coord(0, 32),
+                      offset=WorldCoord(0, 32),
                       size=32,
                       anchor="mm")
     ]
@@ -713,7 +714,7 @@ def main():
     bank[0:5] = [
         CS.point_image(file=Path("bank.png")),
         CS.point_text(colour=0x000000,
-                      offset=Coord(0, 32),
+                      offset=WorldCoord(0, 32),
                       size=32,
                       anchor="mm")
     ]
@@ -723,7 +724,7 @@ def main():
     gym[0:5] = [
         CS.point_image(file=Path("gym.png")),
         CS.point_text(colour=0x000000,
-                      offset=Coord(0, 32),
+                      offset=WorldCoord(0, 32),
                       size=32,
                       anchor="mm")
     ]
@@ -733,7 +734,7 @@ def main():
     shelter[0:5] = [
         CS.point_image(file=Path("shelter.png")),
         CS.point_text(colour=0x000000,
-                      offset=Coord(0, 32),
+                      offset=WorldCoord(0, 32),
                       size=32,
                       anchor="mm")
     ]
@@ -743,7 +744,7 @@ def main():
     playground[0:5] = [
         CS.point_image(file=Path("playground.png")),
         CS.point_text(colour=0x000000,
-                      offset=Coord(0, 32),
+                      offset=WorldCoord(0, 32),
                       size=32,
                       anchor="mm")
     ]
@@ -753,7 +754,7 @@ def main():
     fountain[0:5] = [
         CS.point_image(file=Path("fountain.png")),
         CS.point_text(colour=0x000000,
-                      offset=Coord(0, 32),
+                      offset=WorldCoord(0, 32),
                       size=32,
                       anchor="mm")
     ]
@@ -763,7 +764,7 @@ def main():
     taxi_stand[0:5] = [
         CS.point_image(file=Path("taxistand.png")),
         CS.point_text(colour=0x000000,
-                      offset=Coord(0, 32),
+                      offset=WorldCoord(0, 32),
                       size=32,
                       anchor="mm")
     ]
@@ -773,7 +774,7 @@ def main():
     pick_up_drop_off[0:5] = [
         CS.point_image(file=Path("pickupdropoff.png")),
         CS.point_text(colour=0x000000,
-                      offset=Coord(0, 32),
+                      offset=WorldCoord(0, 32),
                       size=32,
                       anchor="mm")
     ]
@@ -784,7 +785,7 @@ def main():
         bus_stop[i] = [
             CS.point_square(colour=0x00AAFF, size=sq_size, width=int(2 / 3 * sq_size)),
             CS.point_text(
-                colour=0x00AAFF, offset=Coord(0, 20), size=text_size, anchor="mm"
+                colour=0x00AAFF, offset=WorldCoord(0, 20), size=text_size, anchor="mm"
             ),
         ]
     bus_stop[2:3] = [
@@ -797,7 +798,7 @@ def main():
         ferry_stop[i] = [
             CS.point_square(colour=0x1E85AE, size=sq_size, width=int(2 / 3 * sq_size)),
             CS.point_text(
-                colour=0x1E85AE, offset=Coord(0, 20), size=text_size, anchor="mm"
+                colour=0x1E85AE, offset=WorldCoord(0, 20), size=text_size, anchor="mm"
             ),
         ]
     ferry_stop[2:3] = [
@@ -810,7 +811,7 @@ def main():
         rail_station[i] = [
             CS.point_square(colour=0x1F3D7A, size=sq_size, width=int(2 / 3 * sq_size)),
             CS.point_text(
-                colour=0x1F3D7A, offset=Coord(0, 20), size=text_size, anchor="mm"
+                colour=0x1F3D7A, offset=WorldCoord(0, 20), size=text_size, anchor="mm"
             ),
         ]
     rail_station[3:4] = [
@@ -822,7 +823,7 @@ def main():
     transport_exit[0:5] = [
         CS.point_image(file=Path("transportexit.png")),
         CS.point_text(colour=0x000000,
-                      offset=Coord(0, 32),
+                      offset=WorldCoord(0, 32),
                       size=32,
                       anchor="mm"
     ]
@@ -832,7 +833,7 @@ def main():
     attraction[0:5] = [
         CS.point_image(file=Path("attraction.png")),
         CS.point_text(colour=0x000000,
-                      offset=Coord(0, 32),
+                      offset=WorldCoord(0, 32),
                       size=32,
                       anchor="mm"
     ]
@@ -950,7 +951,7 @@ def main():
     simple_point = CTI("point")
     simple_point[0:5] = [
         CS.point_circle(colour=0xFFFFFF, outline=0x000000, size=30, width=6),
-        CS.point_text(colour=0x000000, offset=Coord(10, 10), size=25),
+        CS.point_text(colour=0x000000, offset=WorldCoord(10, 10), size=25),
     ]
     s["simplePoint"] = simple_point