+++ renderer/types/skinbuilder.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from pathlib import Path
 from typing import Literal
 
-from renderer.types import Coord
+from renderer.types.coord import WorldCoord
 
 
 def _hex_to_colour(h: int | None) -> str | None:
@@ -191,7 +191,7 @@ class SkinBuilder:
                 cls,
                 *,
                 colour: int | None = None,
-                offset: Coord = Coord(0, 0),
+                offset: WorldCoord = WorldCoord(0, 0),
                 size: int = 10,
                 anchor: str | None = None,
             ):
@@ -225,7 +225,7 @@ class SkinBuilder:
                 return cs
 
             @classmethod
-            def point_image(cls, *, file: Path, offset: Coord = Coord(0, 0)):
+            def point_image(cls, *, file: Path, offset: WorldCoord = WorldCoord(0, 0)):
                 cs = cls()
                 cs.json = {"layer": "image", "file": str(file), "offset": offset}
                 return cs
@@ -302,7 +302,7 @@ class SkinBuilder:
                 *,
                 colour: int | None = None,
                 size: int = 1,
-                offset: Coord = Coord(0, 0),
+                offset: WorldCoord = WorldCoord(0, 0),
             ):
                 cs = cls()
                 cs.json = {
@@ -331,7 +331,7 @@ class SkinBuilder:
                 return cs
 
             @classmethod
-            def area_centerimage(cls, *, file: Path, offset: Coord = Coord(0, 0)):
+            def area_centerimage(cls, *, file: Path, offset: WorldCoord = WorldCoord(0, 0)):
                 cs = cls()
                 cs.json = {"layer": "centerimage", "file": str(file), "offset": offset}
                 return cs